+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 版本更新 
 
 ## v0.2
+
+### 2021/07/14
+- update `picker component`
+
 ### 2021/06/20
 - add `picker component`